+++ include/OcrUtils.h
@@ -94,6 +94,6 @@ std::string getResultTxtFilePath(const char *path, const char *imgName);
 
 std::string getResultImgFilePath(const char *path, const char *imgName);
 
-std::string getDebugImgFilePath(const char *path, const char *imgName, int i, const char *tag);
+std::string getDebugImgFilePath(const char *path, const char *imgName, size_t i, const char *tag);
 
 #endif //__OCR_UTILS_H__
